+++ src/containers/Organization/Onboarding/Steps/PlatformDetails.tsx
