--- src/containers/Organization/Registration/Registration.module.css
@@ -1,52 +1,0 @@
-.Tooltip {
-  width: 143px;
-  height: 48px;
-  background-color: #042c19 !important;
-  font-size: 10px;
-  line-height: 1.2;
-  border-radius: 8px;
-  font-weight: 400;
-}
-
-.HelperLink {
-  margin-top: 4px !important;
-  text-align: right;
-  color: #93a29b !important;
-  line-height: 1 !important;
-  font-size: 14px !important;
-  font-weight: normal !important;
-  padding-right: 16px;
-}
-
-.InputAdornment {
-  cursor: pointer !important;
-}
-
-.Wrapper :global(.MuiFormControlLabel-root) {
-  display: flex;
-  align-items: flex-start;
-}
-
-.Wrapper :global(.MuiFormControlLabel-label) {
-  color: #93a29b !important;
-  padding: 10px 0px;
-  line-height: 1.2;
-  font-size: 14px;
-  font-weight: 400 !important;
-}
-
-.Wrapper :global(.MuiIconButton-label) {
-  position: relative;
-  z-index: 0;
-}
-
-.Wrapper :global(.MuiIconButton-label:after) {
-  content: '';
-  left: 4px;
-  top: 4px;
-  height: 15px;
-  width: 15px;
-  position: absolute;
-  background-color: white;
-  z-index: -1;
-}
